# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Krazy Meals",
    "description": "Krazy Meals for Swagger",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/products": {
      "post": {
        "tags": [
          "products"
        ],
        "description": "add products",
        "operationId": "post",
        "type": "object",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "add products",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "create response",
            "schema": {
              "$ref": "#/definitions/productResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "products"
        ],
        "description": "what this API does",
        "operationId": "get",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/definitions/get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "create response",
            "schema": {
              "$ref": "#/definitions/response"
            }
          }
        }
      }
    },
    "/api/products/{ID}": {
      "delete": {
        "tags": [
          "products"
        ],
        "description": "Deletes a something",
        "operationId": "delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "description": "ID to delete",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "delete response",
            "schema": {
              "$ref": "#/definitions/deleteRes"
            }
          }
        }
      },
      "get": {
        "tags": [
          "products"
        ],
        "description": "Returns the status of something",
        "operationId": "findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "description": "ID to fetch",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/definitions/response1"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "productResponse": {
      "properties": {
        "active": {
          "type": "boolean",
          "example": true
        },
        "newOld": {
          "type": "string",
          "example": "New"
        },
        "description": {
          "type": "asdasdasd",
          "example": "New"
        },
        "externalProductId": {
          "type": "string",
          "example": "asd12456"
        },
        "model": {
          "type": "string",
          "example": "as321231"
        },
        "make": {
          "type": "string",
          "example": "xyz"
        },
        "productSiteLaunchDate": {
          "type": "date",
          "example": "12-11-2016"
        },
        "merchantReleaseDate": {
          "type": "date",
          "example": "12-11-2016"
        },
        "restockDate": {
          "type": "string",
          "example": "12-11-2016"
        },
        "quantity": {
          "type": "number",
          "example": 10
        },
        "saleFromDate": {
          "type": "date",
          "example": "12-11-2016"
        },
        "saleEndDate": {
          "type": "date",
          "example": "12-11-2016"
        },
        "maxShipQty": {
          "type": "number",
          "example": 2
        },
        "itemPackageQty": {
          "type": "number",
          "example": "1"
        },
        "numberOfItems": {
          "type": "number",
          "example": "1"
        },
        "isGiftwrapped": {
          "type": "boolean",
          "example": false
        },
        "isDiscontinuedByManufacturer": {
          "type": "boolean",
          "example": false
        },
        "isGiftMessaged": {
          "type": "boolean",
          "example": false
        },
        "shippingWeight": {
          "type": "number",
          "example": 100
        },
        "shippingWeightUnitOfMeasure": {
          "type": "string",
          "example": "gm"
        },
        "weightUnitOfMeasure": {
          "type": "string",
          "example": "gm"
        },
        "length": {
          "type": "number",
          "example": 100
        },
        "width": {
          "type": "number",
          "example": 100
        },
        "height": {
          "type": "number",
          "example": "50"
        },
        "point1": {
          "type": "string",
          "example": "boolet point 1"
        },
        "point2": {
          "type": "string",
          "example": "boolet point 2"
        },
        "point3": {
          "type": "string",
          "example": "boolet point 3"
        },
        "point4": {
          "type": "string",
          "example": "boolet point 4"
        },
        "point5": {
          "type": "string",
          "example": "boolet point 5"
        },
        "s3": {
          "type": "string",
          "example": "New"
        },
        "s4": {
          "type": "string",
          "example": "search Keywords 4"
        },
        "s5": {
          "type": "string",
          "example": "search Keywords 5"
        },
        "p3": {
          "type": "string",
          "example": "platinum search Keywords 3"
        },
        "p4": {
          "type": "string",
          "example": "platinum search Keywords 4"
        },
        "s6": {
          "type": "string",
          "example": "search Keywords 6"
        },
        "otherImageUrl": {
          "type": "string",
          "example": "abcd.jpg"
        },
        "otherImageUrl1": {
          "type": "string",
          "example": "abcd1.jpg"
        },
        "otherImageUrl2": {
          "type": "string",
          "example": "abcd2.jpg"
        },
        "otherImageUrl3": {
          "type": "string",
          "example": "abcd3.jpg"
        },
        "otherImageUrl4": {
          "type": "string",
          "example": "abcd4.jpg"
        },
        "otherImageUrl5": {
          "type": "string",
          "example": "abcd5.jpg"
        },
        "otherImageUrl6": {
          "type": "string",
          "example": "abcd6.jpg"
        },
        "otherImageUrl7": {
          "type": "string",
          "example": "abcd7.jpg"
        },
        "otherImageUrl8": {
          "type": "string",
          "example": "abcd8.jpg"
        },
        "color": {
          "type": "string",
          "example": "black"
        },
        "size": {
          "type": "string",
          "example": "xl"
        },
        "sizeMap": {
          "type": "string",
          "example": "xl"
        },
        "colorMap": {
          "type": "string",
          "example": "black"
        },
        "discount": {
          "type": "number",
          "example": 0
        },
        "finalDiscount": {
          "type": "number",
          "example": 0
        },
        "offerAmt": {
          "type": "number",
          "example": 0
        },
        "offerPrice": {
          "type": "number",
          "example": 0
        },
        "discountOnDiscount": {
          "type": "number",
          "example": 0
        },
        "brand": {
          "type": "string",
          "example": "abc"
        },
        "listPrice": {
          "type": "number",
          "example": 500
        },
        "mainImageUrl": {
          "type": "string",
          "example": "abc.jpg"
        },
        "name": {
          "type": "string",
          "example": "test"
        },
        "p1": {
          "type": "string",
          "example": "platinum search Keywords 1"
        },
        "s1": {
          "type": "string",
          "example": "search Keywords 1"
        },
        "s2": {
          "type": "string",
          "example": "search Keywords 2"
        },
        "p2": {
          "type": "string",
          "example": "platinum search Keywords 2"
        },
        "salePrice": {
          "type": "Number",
          "example": 300
        },
        "sku": {
          "type": "string",
          "example": "abc1234"
        },
        "dept": {
          "type": "string",
          "example": "Mexican Grilled Chicken"
        },
        "cat": {
          "type": "food",
          "example": "New"
        },
        "vendorID": {
          "type": "Number",
          "example": 1000
        },
        "instock": {
          "type": "boolean",
          "example": true
        },
        "subCat": {
          "type": "string",
          "example": "meals"
        }
      }
    },
...
    "createProduct": {
      "properties": {
        "name": {
            "required": true,
          "type": "string",
          "example": "chicken"
        },
        "description": {
            "required": true,
          "type": "string",
          "example": "Mexican Grilled Chicken "
        },
        "dept": {
            "required": true,
          "type": "string",
          "example": "non veg"
        },
        "cat": {
            "required": true,
          "type": "string",
          "example": "non veg"
        },
        "subCat": {
            "required": true,
          "type": "string",
          "example": "chicken"
        },
        "sku": {
            "required": true,
          "type": "string",
          "example": "abcd12345"
        },
        "brand": {
          "required": true,
          "type": "string",
          "example": "abcd"
        },

        "s1": {
          "required": true,
          "type": "string",
          "example": "abcd"
        },

        "s2": {
          "required": true,
          "type": "string",
          "example": "abcd"
        },

        "s3": {
          "required": true,
          "type": "string",
          "example": "abcd"
        },
        "p1": {
          "required": true,
          "type": "string",
          "example": "abcd"
        },

        "p2": {
          "required": true,
          "type": "string",
          "example": "abcd"
        },
        "mainImageUrl": {
          "required": true,
          "type": "string",
          "example": "abcd.jpg"
        },
        "vendorID": {
          "required": true,
          "type": "Number",
          "example": 100
        },
        "quantity": {
          "required": true,
          "type": "Number",
          "example": 10
        },
        "listPrice": {
          "required": true,
          "type": "Number",
          "example": 500
        },
        "salePrice": {
          "required": true,
          "type": "Number",
          "example": 300
        },
        "isActive": {
          "required": true,
          "type": "boolean",
          "example": true
        }
      }
    },
...
    "response1": {
      "properties": {
        "Id": {
          "required": true,
          "type": "Number",
          "example": "123"
        }
      }
    },
...
    "response": {
      "properties": {
        "Id": {
          "required": true,
          "type": "Number",
          "example": "12"
        }
      }
    },
...
  }
}